+++ src/main/java/com/webtooni/webtooniverse/domain/talktalk/dto/requset/TalkLikeRequestDto.java
@@ -1,19 +1,18 @@
-package com.webtooni.webtooniverse.domain.talktalk.dto;
+package com.webtooni.webtooniverse.domain.talktalk.dto.requset;
 
 import com.webtooni.webtooniverse.domain.talktalk.domain.TalkPost;
 import com.webtooni.webtooniverse.domain.user.domain.User;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
-public class TalkReviewRequestDto {
+@NoArgsConstructor
+public class TalkLikeRequestDto {
 
     /**
      * TODO 사용여부 확인
      */
-
     private TalkPost talkPost;
     private User user;
-    private String commentContent;
-
 
 }
