+++ src/main/java/com/webtooni/webtooniverse/domain/talktalk/repository/TalkPostRepository.java
@@ -4,4 +4,5 @@ import com.webtooni.webtooniverse.domain.talktalk.domain.TalkPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TalkPostRepository extends JpaRepository<TalkPost, Long> {
+
 }
